refactor : 도메인의 패키지를 변경한다
