+++ demo/src/main/java/com/example/demo/order/service/OrderService.java
@@ -1,0 +1,11 @@
+package com.example.demo.order.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class OrderService {
+
+    //@Autowired
+    //private OrderRepository orderRepository;
+}
